refactor: implement custom render in home page
